requirements.txt
langgraph
langchain-core
langchain-community
langchain-openai
langchain-anthropic
pygithub
atlassian-python-api
python-dotenv
slack-bolt